requirements.txt
aiosmtplib==1.1.7
allure-pytest==2.12.0
allure-python-commons==2.12.0
anyio==3.7.1
apturl==0.5.2
argcomplete==3.5.0
async-generator==1.10
attrs==22.2.0
autopep8==2.0.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.0.1
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.6
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info==0.23+ubuntu1.1
duplicity==0.8.12.0
entrypoints==0.3
exceptiongroup==1.2.2
Faker==16.8.1
fasteners==0.14.1
future==0.18.2
h11==0.14.0
httpcore==0.13.7
httplib2==0.14.0
httpx==0.19.0
idna==3.4
iniconfig==2.0.0
iotop==0.6
keyboard==0.13.5
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libvirt-python==6.1.0
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
messages==0.8.0
monotonic==1.5
MouseInfo==0.1.3
netifaces==0.10.4
numpy==1.24.4
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
outcome==1.2.0
packaging==23.0
pandas==2.0.3
paramiko==2.6.0
pexpect==4.6.0
pillow==10.4.0
pipx==1.7.1
platformdirs==4.3.6
pluggy==1.0.0
protobuf==3.6.1
PyAutoGUI==0.9.54
pycairo==1.16.2
pycodestyle==2.10.0
pycups==1.9.73
PyGetWindow==0.0.9
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMsgBox==2.0.1
PyNaCl==1.3.0
pyparsing==2.4.6
pyperclip==1.11.0
PyQt5==5.14.1
PyRect==0.2.0
pyRFC3339==1.1
PyScreeze==1.0.1
PySocks==1.7.1
pyTelegramBotAPI==4.22.0
pytest==7.2.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-dotenv==0.21.1
python-xlib==0.33
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.2
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.28.2
requests-unixsocket==0.2.0
rfc3986==1.5.0
SecretStorage==2.3.1
selenium==4.8.0
simplejson==3.16.0
sip==4.19.21
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
ssh-import-id==5.10
systemd-python==234
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.14
usb-creator==0.3.7
userpath==1.9.2
validus==0.3.0
wadllib==1.3.3
webdriver-manager==3.8.5
wsproto==1.2.0
xkit==0.0.0
